// src/types/business.ts
export interface BusinessFormValues {
  businessName: string;
  ownerName: string;
  email: string;
  phone: string;
  address: string;
  businessType: string;
  description: string;
  logo_url: string;
  password: string;
  confirmPassword: string;
  subscriptionPlan: string;
  preapproval_id: string;
}